Fixes EF database issue with casting decimal to double (schema errors). Adds support to inventory page to detect resistor/capacitor values when adding. Added better package type UI. Added quick toggle for enabling/disabling auto part info search. Fixed CSS issues.

+++ Binner/Data/Binner.Data.Model/PartParametric.cs
@@ -31,7 +31,7 @@ namespace Binner.Data.Model
         /// <summary>
         /// Value as a number if numeric
         /// </summary>
-        public double ValueNumber { get; set; }
+        public decimal ValueNumber { get; set; }
 
         /// <summary>
         /// The measurement units of the value
